requirements.txt
discord.py[voice]
pynacl
asyncio
psutil
requests
wikipedia
datetime
pymongo
dnspython
google
uptime
imagescraper